refactor(spa): extract GeneratorComponent from monolith (T4b) [KAN-118]
Move the entire recipe generation view (~565 lines template, ~260 lines
TS) into its own component. GeneratorComponent injects services directly
and owns all generation state (prompt, loading, error, scaling, notes
editing, publish controls). Emits addToCookbookRequested and
authModalRequested outputs for modals still in AppComponent.

AppComponent: 931→673 TS, 1824→1267 HTML (1940 total, was 2755).
